Allow 'ls' to be written to bash history
Along with preexec command, we show the last run command at the top of
the screen, with its output below. When we erased ls from the history,
it prevented it from being shown as the 'last' command, and instead
showed whatever was executed before it. I'm adding it back in so it is
properly shown again.